Load authentication component on-demand
