+++ website/package.json
@@ -20,7 +20,7 @@
     "@fontsource/charis-sil": "^4.4.2",
     "@fontsource/quattrocento-sans": "^4.4.2",
     "@fortawesome/fontawesome-free": "^5.15.1",
-    "@loadable/component": "^5.14.1",
+    "@loadable/component": "^5.15.0",
     "@reach/router": "^1.3.4",
     "@react-hook/debounce": "^3.0.0",
     "aws-amplify": "^4.0.2",
@@ -57,6 +57,7 @@
     "@babel/preset-typescript": "^7.13.0",
     "@types/color": "^3.0.1",
     "@types/js-cookie": "^2.2.6",
+    "@types/loadable__component": "^5.13.3",
     "@types/lodash": "^4.14.161",
     "@types/node": "^15.6.2",
     "@types/pluralize": "^0.0.29",
@@ -105,6 +106,7 @@
     "graphql": "^15.4.0",
     "graphql-compose": "^7.25.0",
     "webpack": "^5.24.2",
-    "@babel/core": "^7.14.3"
+    "@babel/core": "^7.14.3",
+    "@babel/runtime": "^7.14.0"
   }
 }
